# improve type inference and speedup norm