--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -3,40 +3,23 @@ using Test
 import LinearAlgebra.SingularException
 
 @testset "CliffordAlgebras.jl" begin
-    @testset "inference" begin
-        alg = CliffordAlgebra(1,1,1)
-        @inferred 3*alg.e1
-        @inferred 3.0*alg.e1
-        @inferred alg.e1*3f0
-        @inferred (3alg.e1) * (alg.e2)
-        @inferred alg.e1 ∧ 2.3
-    
-        @inferred dual(alg.e1)
-        @inferred dual(alg.e1 + alg.e1e2)
-        @inferred alg.e1 ∨ (alg.e2e3 + 1)
-    
-        @inferred scalar(alg.e1)
-        @inferred scalar(3*alg.e1)
-        @inferred reverse(alg.e1 + 3*alg.e1e2)
-        @inferred norm(alg.e1 + 3*alg.e1e2)
+    @testset "inference pga3d" begin
+        pga = CliffordAlgebra(3,0,1)
+        pt  = @inferred dual(pga.e4 + 3.2pga.e1 + 1.3pga.e2-4.3pga.e3)
+        pt1 = @inferred dual(pga.e4 + pga.e1)
+        pt2 = @inferred dual(pga.e4 + pga.e2 - pga.e3)
+        l = @inferred ∨(pt1, pt2)
+        @inferred norm(l)
+        l = @inferred l / norm(l)
+        motor1 = @inferred exp(-pi/6*l)
+        motor2 = @inferred exp(-1/2*pga.e3e4)
+        motor = @inferred motor1 * motor2
 
-        @testset "inference pga3d" begin
-            pga = CliffordAlgebra(3,0,1)
-            pt  = @inferred dual(pga.e4 + 3.2pga.e1 + 1.3pga.e2-4.3pga.e3)
-            pt1 = @inferred dual(pga.e4 + pga.e1)
-            pt2 = @inferred dual(pga.e4 + pga.e2 - pga.e3)
-            l = @inferred ∨(pt1, pt2)
-            l = @inferred l / norm(l)
-            motor1 = @inferred exp(-pi/6*l)
-            motor2 = @inferred exp(-1/2*pga.e3e4)
-            motor = @inferred motor1 * motor2
-
-            @inferred ≀(motor, pt)
-            @inferred ≀(motor1, pt)
-            @inferred ≀(motor2, pt)
-            @inferred ≀(motor, l)
-            @inferred ≀(motor, motor1)
-        end
+        @inferred ≀(motor, pt)
+        @inferred ≀(motor1, pt)
+        @inferred ≀(motor2, pt)
+        @inferred ≀(motor, l)
+        @inferred ≀(motor, motor1)
     end
 
     @testset "isapprox" begin
@@ -210,6 +193,24 @@ import LinearAlgebra.SingularException
         s1 = scalar(e1*e1)
         s2 = scalar(e2*e2)
         s3 = scalar(e3*e3)
+
+        @testset "inference" begin
+            @inferred 3*e1
+            @inferred 3.0*e1
+            @inferred e1*3f0
+            @inferred (3e1) * (e2)
+            @inferred e1 ∧ 2.3
+            @inferred -e1
+        
+            @inferred dual(e1)
+            @inferred dual(e1 + e12)
+            @inferred e1 ∨ (e23 + 1)
+        
+            @inferred scalar(e1)
+            @inferred scalar(3*e1)
+            @inferred reverse(e1 + 3*e12)
+            @inferred norm(e1 + 3*e12)
+        end
 
         @testset "addition/subtraction" begin
             @test e1 + e2 == e2 + e1