Fixes various inference and performance issues. The most extreme one:
```julia
using CliffordAlgebras
using BenchmarkTools

pga = CliffordAlgebra(3,0,1)

mv = dual(pga.e4 + 3.2pga.e1 + 1.3pga.e2-4.3pga.e3)
 
@btime norm($mv)

# before:  74.091 ns (4 allocations: 96 bytes)

# after:   1.560 ns (0 allocations: 0 bytes)
```

In an actual use case that involves 3d PGA this gives me a nice speedup:
```
# Before:
5.78 seconds 4.9GiB allocs
# After:
0.072 seconds 1.679 MiB allocs
```