@ATell-SoundTheory 2023-01-11
Thank you so much for all the work! I'v been meaning to look into exactly these things myself, but had not found the time. The changes all make sense, but I'm slightly surprised that moving an inner constructor to the outside has any kind of performance impact. Maybe I've missed the point, as I've only briefly skimmed over the changes.

I'll merge the PR when I find the time to do a bit of a more in-depth review. I have also planned to make a new release soon with all the accumulated improvements.