@jw3126 2023-01-11
> The changes all make sense, but I'm slightly surprised that moving an inner constructor to the outside has any kind of performance impact.

Ok I did not explain this very well. First of all, simply moving a constructor out does not have any performance impact. However moving the inner constructor out gives us the option to not use it. Instead of
```julia
MultiVector(CA, BI, coeffs)
```
we can directly call
```julia
MultiVector{CA, T, BI, K}(coeffs)
```
The latter is trivial simple to infer, while the former has a subtle issue.
The issue is, that the former constructor is not type stable. `BI = (1,2)` vs `BI = (2,3)` will give different return types, for the same input type.

Generally speaking, julia does an awesome job at inferring type stable code.
Code that requires constant propagation to figure out the type does not infer as well. Although
const prop gets better with every release. In most places the old inner constructor worked for me on julia 1.9, while it failed on julia 1.6.
Generally, my strategy is to keep things simple stupid and avoid reliance on fancy compiler heuristics, like const prop.