@jw3126 2023-01-11
> I'll merge the PR when I find the time to do a bit of a more in-depth review. I have also planned to make a new release soon with all the accumulated improvements.

Awesome I wanted to ask you about a release after this is merged anyway.